requirements.txt
streamlit
transformers
huggingface-hub
google-generativeai
python-dotenv
